add animateGrid()
