+++ awatif-ui/src/viewer/drawing/animateGrid/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Awatif UI - Drawing - Animate Grid</title>
+    <script type="module">
+      import { viewer } from "../../viewer";
+      import van from "vanjs-core";
+
+      const gridTarget = van.state({
+        position: [10, 0, 0],
+        rotation: [Math.PI / 2, Math.PI / 2, Math.PI / 2],
+      });
+
+      const viewerElm = viewer({
+        drawingObj: {
+          gridTarget,
+        },
+      });
+
+      document.body.appendChild(viewerElm);
+    </script>
+  </head>
+  <body></body>
+</html>
