+++ awatif-ui/src/viewer/drawing/drawPoints/index.html
@@ -3,9 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Awatif UI - Drawing</title>
+    <title>Awatif UI - Drawing - Draw Points</title>
     <script type="module">
-      import { viewer } from "../viewer";
+      import { viewer } from "../../viewer";
       import van from "vanjs-core";
 
       const points = van.state([[10, 0, 0]]);
